added watermarking script
the script won't be functional because the paths in it are absolute. even if they weren't, it assumes the user has a folder called "images" in the same directory, which itself contains two directories named "watermark" and "no-watermark". the images are taken from the "no-watermark" folder and saved to the "watermark" folder.